chore(kuma-cp) split generators into separate files
